broke script apart into smaller modules
