+++ config/style.config.js
@@ -1,4 +1,5 @@
 
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports.entry = ()=>({
     style: './src/scss/styles.scss',
