+++ gardner_anon.csv
@@ -1,25 +1,25 @@
-name,paygrade,station,duty,start_date,end_date,months,rs_name,rs_paygrade,rs_title,trait_1,trait_2,trait_3,trait_4,trait_5,trait_avg,rsca,prom_rec,n_sp,n_pr,n_p,n_mp,n_ep,prt,rpt_type
-"GARDNER, THOMAS WILLIAM II",O1,NAVNUPWRTRACOM,STUDENT,2005-10-21,2006-06-12,8,NAVY J J,O5,SCHOOL DIR,,,,,,,,nob,0,0,0,0,0,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O1,NAVNUPWRTRAUCHASN,STUDENT,2006-06-17,2006-12-15,6,CAPTAIN A A,O5,OIC,,,,,,,,nob,0,0,0,0,0,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O1,SSN 723 OKLAHOMA C,SONAR DIV OFF,2006-12-16,2007-05-31,6,SUB C O,O5,CO,,,,,,,,nob,0,0,0,0,0,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O2,SSN 723 OKLAHOMA C,SONAR,2007-06-01,2008-02-29,9,SUB C O,O5,CO,0,0,3,3,0,3.5,4.04,p,0,0,8,0,0,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O2,SSN 723 OKLAHOMA C,CRA,2008-03-01,2008-11-22,9,SUB C O,O5,CO,0,0,0,0,7,5,4.03,p,0,0,5,0,0,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O2,SSN 723 OKLAHOMA C,CRA,2008-11-23,2009-02-28,3,SAILOR I M,O5,CO,0,0,0,3,4,4.57142857142857,3.75,p,0,0,6,0,0,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O3,SSN 723 OKLAHOMA C,CRA,2009-03-01,2010-01-31,11,SAILOR I M,O5,CO,0,0,0,5,1,4.16666666666667,4.11,mp,0,0,0,3,1,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O3,SSN 723 OKLAHOMA C,AENG,2010-02-01,2010-04-02,2,SAILOR I M,O5,CO,0,0,0,4,2,4.33333333333333,4.12,ep,0,0,0,0,1,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O3,TASK FORCE 69,WATCH OFFICER,2010-04-03,2010-07-15,3,COMMODORE Z Z,O6,COMMANDER,0,0,6,0,0,3,4.12,mp,0,0,0,12,4,P/WS,RG
-"GARDNER, THOMAS WILLIAM II",O3,CNE-CNA-C6F,DIV OFFICER,2010-07-16,2011-01-31,7,AVIATOR C O,O6,DIRECTOR COPS,0,0,0,6,0,4,4.33,mp,0,0,0,13,3,P,RG
-"GARDNER, THOMAS WILLIAM II",O3,CNE-CNA-C6F,MOC DIVO,2011-02-01,2012-01-01,12,FLORES K T,O6,DIRECTOR COPS,0,0,1,3,2,4.16666666666667,4.17,mp,0,0,5,5,3,PP,RG
-"GARDNER, THOMAS WILLIAM II",O3,CNE-CNA-C6F,MOC DEPUTY DIR,2012-02-01,2013-01-31,12,FLORES K T,O6,DIRECTOR COPS,0,0,0,4,2,4.33333333333333,4.2,ep,0,0,0,0,1,PP,RG
-"GARDNER, THOMAS WILLIAM II",O3,CNE-CNA-C6F,MOC DEPUTY DIR,2013-02-01,2013-03-31,4,FLORES K T,O6,DIRECTOR N,0,0,0,0,6,5,4.2,ep,0,0,0,0,1,N,RG
-"GARDNER, THOMAS WILLIAM II",O3,NR UWO DET L,ADMIN DIV OFF,2013-06-01,2014-01-31,8,JONES A A,O6,CO,0,0,0,5,1,4.16666666666667,4.09,ep,0,0,3,3,2,P,IDT/RG
-"GARDNER, THOMAS WILLIAM II",O3,NR UWO DET L,WATCH OFFICER,2014-02-01,2015-01-31,12,JONES A A,O6,CO,0,0,0,4,2,4.33333333333333,4.07,ep,0,0,1,1,1,PP,IDT/RG
-"GARDNER, THOMAS WILLIAM II",O4,NR UWO DET L,OPS DH,2015-02-01,2015-10-31,9,SMITH B B,O6,CO,0,0,0,5,1,4.16666666666667,3.98,mp,0,0,5,3,3,PP,IDT/RG
-"GARDNER, THOMAS WILLIAM II",O4,SUBLANT CTF-84 UWO,OPS DH,2015-11-01,2016-06-30,8,SMITH B B,O6,CO,0,0,0,4,2,4.33333333333333,4.01,ep,0,0,0,0,1,P,IDT/RG
-"GARDNER, THOMAS WILLIAM II",O4,NR SUBLANT UWO DC,ADMIN DH,2016-11-01,2016-10-31,4,JOHNSON C C,O5,CO,0,0,2,2,2,4,4.03,p,0,0,2,1,1,P,IDT/RG
-"GARDNER, THOMAS WILLIAM II",O4,NR SUBLANT UWO DC,N1 DEPT HEAD,2016-11-01,2017-06-15,8,JOHNSON C C,O5,CO,0,0,1,3,2,4.16666666666667,4.04,ep,0,0,0,0,1,P,IDT/RG
-"GARDNER, THOMAS WILLIAM II",O4,SOCFWD NWA,J5 PLANS DIR,2017-06-16,2017-10-31,5,ARMY GUY,O6,COMMANDER,0,0,2,4,1,3.85714285714286,4.04,ep,0,0,0,0,1,N,AT/RG
-"GARDNER, THOMAS WILLIAM II",O4,SOCFWD NWA,J8 RES MGT DIR,2017-11-01,2018-06-01,7,ARMY GUY2,O6,COMMANDER,0,0,1,3,2,4.16666666666667,4.17,ep,0,0,0,0,1,PP,AT/RG
-"GARDNER, THOMAS WILLIAM II",O4,NR SUBLANT UWO DC,N8 DEPT HEAD,2018-06-02,2018-10-31,5,BROWN D D,O6,CO,0,0,0,4,2,4.33333333333333,4.05,mp,0,0,4,2,2,P,IDT/RG
-"GARDNER, THOMAS WILLIAM II",O4,NR SUBLANT UWO DC,DEPT HEAD,2018-11-01,2019-11-30,13,BROWN D D,O6,CO,0,0,0,3,3,4.5,4.08,p,0,0,2,1,1,BP,IDT/RG
-"GARDNER, THOMAS WILLIAM II",O5,NR CNO CAPS ASMNT,ACTION OFFICER,2019-12-01,2020-08-31,9,ANOTHER C O,O6,CO,0,0,1,4,1,4,3.90,ep,0,0,0,0,1,N,IDT/RG
+paygrade,station,duty,start_date,end_date,months,rs_name,rs_paygrade,rs_title,trait_1,trait_2,trait_3,trait_4,trait_5,trait_avg,rsca,prom_rec,n_sp,n_pr,n_p,n_mp,n_ep,prt,rpt_type
+O1,NAVNUPWRTRACOM,STUDENT,2005-10-21,2006-06-12,8,NAVY J J,O5,SCHOOL DIR,,,,,,,,nob,0,0,0,0,0,P/WS,RG
+O1,NAVNUPWRTRAUCHASN,STUDENT,2006-06-17,2006-12-15,6,CAPTAIN A A,O5,OIC,,,,,,,,nob,0,0,0,0,0,P/WS,RG
+O1,SSN 723 OKLAHOMA C,SONAR DIV OFF,2006-12-16,2007-05-31,6,SUB C O,O5,CO,,,,,,,,nob,0,0,0,0,0,P/WS,RG
+O2,SSN 723 OKLAHOMA C,SONAR,2007-06-01,2008-02-29,9,SUB C O,O5,CO,0,0,3,3,0,3.5,4.04,p,0,0,8,0,0,P/WS,RG
+O2,SSN 723 OKLAHOMA C,CRA,2008-03-01,2008-11-22,9,SUB C O,O5,CO,0,0,0,0,7,5,4.03,p,0,0,5,0,0,P/WS,RG
+O2,SSN 723 OKLAHOMA C,CRA,2008-11-23,2009-02-28,3,SAILOR I M,O5,CO,0,0,0,3,4,4.57142857142857,3.75,p,0,0,6,0,0,P/WS,RG
+O3,SSN 723 OKLAHOMA C,CRA,2009-03-01,2010-01-31,11,SAILOR I M,O5,CO,0,0,0,5,1,4.16666666666667,4.11,mp,0,0,0,3,1,P/WS,RG
+O3,SSN 723 OKLAHOMA C,AENG,2010-02-01,2010-04-02,2,SAILOR I M,O5,CO,0,0,0,4,2,4.33333333333333,4.12,ep,0,0,0,0,1,P/WS,RG
+O3,TASK FORCE 69,WATCH OFFICER,2010-04-03,2010-07-15,3,COMMODORE Z Z,O6,COMMANDER,0,0,6,0,0,3,4.12,mp,0,0,0,12,4,P/WS,RG
+O3,CNE-CNA-C6F,DIV OFFICER,2010-07-16,2011-01-31,7,AVIATOR C O,O6,DIRECTOR COPS,0,0,0,6,0,4,4.33,mp,0,0,0,13,3,P,RG
+O3,CNE-CNA-C6F,MOC DIVO,2011-02-01,2012-01-01,12,FLORES K T,O6,DIRECTOR COPS,0,0,1,3,2,4.16666666666667,4.17,mp,0,0,5,5,3,PP,RG
+O3,CNE-CNA-C6F,MOC DEPUTY DIR,2012-02-01,2013-01-31,12,FLORES K T,O6,DIRECTOR COPS,0,0,0,4,2,4.33333333333333,4.2,ep,0,0,0,0,1,PP,RG
+O3,CNE-CNA-C6F,MOC DEPUTY DIR,2013-02-01,2013-03-31,4,FLORES K T,O6,DIRECTOR N,0,0,0,0,6,5,4.2,ep,0,0,0,0,1,N,RG
+O3,NR UWO DET L,ADMIN DIV OFF,2013-06-01,2014-01-31,8,JONES A A,O6,CO,0,0,0,5,1,4.16666666666667,4.09,ep,0,0,3,3,2,P,IDT/RG
+O3,NR UWO DET L,WATCH OFFICER,2014-02-01,2015-01-31,12,JONES A A,O6,CO,0,0,0,4,2,4.33333333333333,4.07,ep,0,0,1,1,1,PP,IDT/RG
+O4,NR UWO DET L,OPS DH,2015-02-01,2015-10-31,9,SMITH B B,O6,CO,0,0,0,5,1,4.16666666666667,3.98,mp,0,0,5,3,3,PP,IDT/RG
+O4,SUBLANT CTF-84 UWO,OPS DH,2015-11-01,2016-06-30,8,SMITH B B,O6,CO,0,0,0,4,2,4.33333333333333,4.01,ep,0,0,0,0,1,P,IDT/RG
+O4,NR SUBLANT UWO DC,ADMIN DH,2016-11-01,2016-10-31,4,JOHNSON C C,O5,CO,0,0,2,2,2,4,4.03,p,0,0,2,1,1,P,IDT/RG
+O4,NR SUBLANT UWO DC,N1 DEPT HEAD,2016-11-01,2017-06-15,8,JOHNSON C C,O5,CO,0,0,1,3,2,4.16666666666667,4.04,ep,0,0,0,0,1,P,IDT/RG
+O4,SOCFWD NWA,J5 PLANS DIR,2017-06-16,2017-10-31,5,ARMY GUY,O6,COMMANDER,0,0,2,4,1,3.85714285714286,4.04,ep,0,0,0,0,1,N,AT/RG
+O4,SOCFWD NWA,J8 RES MGT DIR,2017-11-01,2018-06-01,7,ARMY GUY2,O6,COMMANDER,0,0,1,3,2,4.16666666666667,4.17,ep,0,0,0,0,1,PP,AT/RG
+O4,NR SUBLANT UWO DC,N8 DEPT HEAD,2018-06-02,2018-10-31,5,BROWN D D,O6,CO,0,0,0,4,2,4.33333333333333,4.05,mp,0,0,4,2,2,P,IDT/RG
+O4,NR SUBLANT UWO DC,DEPT HEAD,2018-11-01,2019-11-30,13,BROWN D D,O6,CO,0,0,0,3,3,4.5,4.08,p,0,0,2,1,1,BP,IDT/RG
+O5,NR CNO CAPS ASMNT,ACTION OFFICER,2019-12-01,2020-08-31,9,ANOTHER C O,O6,CO,0,0,1,4,1,4,3.9,ep,0,0,0,0,1,N,IDT/RG
