+++ index.html
@@ -1,18 +1,13 @@
 <head>
   <meta charset="utf-8" />
   <link rel="stylesheet" href="./dist/style.min.css" />
-  <!-- <script type="text/javascript" src="./dist/app.js"></script> -->
-<!-- <script type="text/javascript" src="./dist/pdfjs.js"></script> -->
+  <script type="text/javascript" src="./dist/app.js"></script>
 </head>
 
 <body>
-  <div id="parser">  
-    <input type="file" name="psr" id="psr" onchange="parseUploadedForm(this)"/>
-  </div>
   <div class="grid">
     <div class="name"><h1 >PSR</h1></div>
     <div class="gh_link" ><h2  ><a href="https://github.com/twgardner2/navy_psr/issues"
       >Click here to leave feedback</a><h2></div>
       </div>
-      <script type="text/javascript" src="./dist/parser.js"></script>
 </body>
