--- src/js/parser/helpers.js
@@ -1,9 +1,0 @@
-
-
-function convertToThou(num){
-    return Math.round(num*1000);
-}
-
-module.exports={
-    convertToThou
-}
